pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kitty"
version = "0.1.0"
description = "한국투자증권 자동 매매 시스템 - 멀티 에이전트 기반"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "python-telegram-bot>=21.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "loguru>=0.7",
    "apscheduler>=3.10",
    "aiohttp>=3.9",
    "pandas>=2.0",
    "numpy>=1.26",
]

[project.optional-dependencies]
openai = [
    "openai>=1.30",
]
gemini = [
    "google-generativeai>=0.7",
]
all-providers = [
    "openai>=1.30",
    "google-generativeai>=0.7",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
kitty = "kitty.main:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
